rename: csr files
